requirements.txt
langchain-groq
langchain-community
langchain-text-splitters
langchain-core
langgraph
chromadb
sentence-transformers
pypdf
python-dotenv
duckduckgo-search
streamlit